chore: Correct data type in all IDS, previous data = int, corrected data = Long, located in: src/main/java/com/juanhiginio/genomeBank/dto

+++ src/main/java/com/juanhiginio/genomeBank/dto/CrearCromosomaInDTO.java
@@ -18,7 +18,7 @@ public class CrearCromosomaInDTO {
     private String secuenciaADN;
 
     @NotNull
-    private int genoma_id;
+    private Long genoma_id;
