+++ src/main/java/com/juanhiginio/genomeBank/dto/CrearCromosomaOutDTO.java
@@ -5,12 +5,12 @@ import lombok.Data;
 @Data
 public class CrearCromosomaOutDTO {
 
-    private int id;
+    private Long id;
     private String nombre;
     private int longitudParesDeBases;
     private String secuenciaADN;
 
     //Relación con genoma y especie
-    private int genoma_id;
+    private Long genoma_id;
     private String nombreCientifico;
 }
